refactor: #35 Modify Appbar related logic
