+++ src/component/common/appBar/index.ts
@@ -1,2 +1,3 @@
 export { AppBar } from "./AppBar";
+export { UserProfileIcon } from "./UserProfileIcon";
 export type { AppBarProps } from "./AppBar";
